@@ -1,4 +1,5 @@
-import { useRouter } from "next/router";
+"use client";
+import { useRouter, usePathname } from "next/navigation";
 import { useEffect, useState } from "react";
 import { Pagination } from "./Pagination";
 import { Results } from "./Results";
@@ -10,6 +11,7 @@ export const PropertySearch = () => {
   const [totalResults, setTotalResults] = useState(0);
   const pageSize = 3;
   const router = useRouter();
+  const pathname = usePathname();
 
   const search = async () => {
     const { page, minPrice, maxPrice, hasParking, petFriendly } =
@@ -36,7 +38,6 @@ export const PropertySearch = () => {
       }),
     });
     const data = await response.json();
-    console.log("SEARCH DATA: ", data);
     setProperties(data.properties);
     setTotalResults(data.total);
   };
@@ -46,18 +47,13 @@ export const PropertySearch = () => {
       window.location.search
     );
 
-    await router.push(
-      `${router.query.slug.join("/")}?page=${pageNumber}&petFriendly=${
+    router.push(
+      `${pathname}?page=${pageNumber}&petFriendly=${
         petFriendly === "true"
       }&hasParking=${
         hasParking === "true"
-      }&minPrice=${minPrice}&maxPrice=${maxPrice}`,
-      null,
-      {
-        shallow: true,
-      }
+      }&minPrice=${minPrice}&maxPrice=${maxPrice}`
     );
-    search();
   };
 
   useEffect(() => {
@@ -72,17 +68,9 @@ export const PropertySearch = () => {
   }) => {
     // update our browser url
     // search
-    console.log("FILTERS: ", petFriendly, hasParking, minPrice, maxPrice);
-    await router.push(
-      `${router.query.slug.join(
-        "/"
-      )}?page=1&petFriendly=${!!petFriendly}&hasParking=${!!hasParking}&minPrice=${minPrice}&maxPrice=${maxPrice}`,
-      null,
-      {
-        shallow: true,
-      }
+    router.push(
+      `${pathname}?page=1&petFriendly=${!!petFriendly}&hasParking=${!!hasParking}&minPrice=${minPrice}&maxPrice=${maxPrice}`
     );
-    search();
   };
 
   return (
