@@ -1,18 +1,15 @@
-import { usePageContext } from "context/page";
 import Image from "next/image";
 
 export const Cover = ({ children, background }) => {
-  const { featuredImage } = usePageContext();
-
   return (
     <div className="h-screen text-white bg-slate-800 relative min-h-[400px] flex justify-center items-center">
-      {(!!background || !!featuredImage) && (
+      {!!background && (
         <Image
           alt="Cover"
-          src={background || featuredImage}
-          layout="fill"
-          objectFit="cover"
-          className="mix-blend-soft-light"
+          src={background}
+          fill
+          priority="low"
+          className="mix-blend-soft-light object-cover"
         />
       )}
       <div className="max-w-5xl z-10">{children}</div>
