@@ -28,26 +28,11 @@ export const Gallery = ({ columns, cropImages, items }) => {
             src={item.attributes.url}
             height={maxHeight || item.attributes.height}
             width={maxWidth || item.attributes.width}
-            alt={item.attributes.alt}
-            objectFit="cover"
+            alt={item.attributes.alt || ""}
+            className="object-cover"
           />
         </div>
       ))}
     </div>
   );
 };
-
-/*let maxHeight = 0;
-  let maxWidth = 0;
-
-  if (cropImages) {
-    items.forEach((item) => {
-      if (item.attributes.height > maxHeight) {
-        maxHeight = item.attributes.height;
-      }
-      if (item.attributes.width > maxWidth) {
-        maxWidth = item.attributes.width;
-      }
-    });
-  }
-*/
