@@ -7,7 +7,6 @@ export const MainMenu = ({
   callToActionLabel,
   callToActionDestination,
 }) => {
-  console.log("MAIN MENU: ", items);
   return (
     <div className="bg-slate-800 text-white px-5 h-[64px] sticky top-0 z-20 flex">
       <div className="py-4 pl-5 flex text-pink-600">
@@ -21,17 +20,19 @@ export const MainMenu = ({
             className="hover:bg-slate-700 cursor-pointer relative group"
           >
             <div>
-              <Link href={item.destination}>
-                <a className="p-5 block">{item.label}</a>
+              <Link href={item.destination} className="p-5 block">
+                {item.label}
               </Link>
             </div>
             {!!item.subMenuItems?.length && (
               <div className="group-hover:block hidden bg-slate-800 text-right absolute right-0 top-full -mt-3">
                 {item.subMenuItems.map((subMenuItem) => (
-                  <Link key={subMenuItem.id} href={subMenuItem.destination}>
-                    <a className="block whitespace-nowrap p-5 hover:bg-slate-700">
-                      {subMenuItem.label}
-                    </a>
+                  <Link
+                    key={subMenuItem.id}
+                    href={subMenuItem.destination}
+                    className="block whitespace-nowrap p-5 hover:bg-slate-700"
+                  >
+                    {subMenuItem.label}
                   </Link>
                 ))}
               </div>
