@@ -1,18 +1,10 @@
 import { BlockRenderer } from "components/BlockRenderer";
 import { MainMenu } from "components/MainMenu";
-import { PageWrapper } from "context/page";
 import Head from "next/head";
 
 export const Page = (props) => {
-  console.log("PAGE PROPS: ", props);
   return (
-    <PageWrapper
-      value={{
-        propertyFeatures: props.propertyFeatures,
-        title: props.title,
-        featuredImage: props.featuredImage,
-      }}
-    >
+    <div>
       <Head>
         <title>{props.seo.title}</title>
         <meta name="description" content={props.seo.metaDesc} />
@@ -23,6 +15,6 @@ export const Page = (props) => {
         callToActionLabel={props.callToActionLabel}
       />
       <BlockRenderer blocks={props.blocks} />
-    </PageWrapper>
+    </div>
   );
 };
